refactor(connection): modular connection/ dir + Dalga 2-3 UX/error/Bytes changes
Why:
1535 satırlık `src/connection.rs` monolit mimari en büyük tek borç
(mimari araştırması raporu). 7 sorumluluk-bazlı alt-modüle bölündü;
public API (pub async fn connection::handle(...)) değişmedi.

Aynı zamanda Dalga 2-3'ün connection.rs'e dokunan UX + error enum
+ full Bytes değişiklikleri de bu PR'da birlikte geliyor (PR #3
sadece 3 satır .into() adaptasyonu içeriyordu):
  * classify_handshake_error — UKEY2 hata sınıflandırması
  * PIN mismatch + timeout + disconnect user-facing notify
  * Legacy → hash trust migration dialog öncesi açıklama
  * bail!() → HekaError variant migration (~15 site)
  * Full Bytes refactor (secret_id_hash, signed_data, body, vs)

## Yeni dizin yapısı

```
src/connection/
├── mod.rs       (384) — pub async fn handle() dispatcher + glue
├── consent.rs   (368) — FlowOutcome enum + handle_sharing_frame
├── sanitize.rs  (352) — sanitize_received_name + path helpers
├── sharing.rs   (283) — text payload + preview + cleanup
├── frames.rs    (190) — build_* / send_* / wrap_payload_transfer
├── request.rs   (131) — ConnectionRequest parse + rate limit
└── errors.rs    (68)  — classify_handshake_error
```

Toplam 1776 satır (öncesi 1535) — artış import/module deklarasyon
+ pub(crate) visibility ve inline testler.

handle() dispatcher: 365 → 332 satır. Encrypted loop body hâlâ
inline; extraction için ConnectionCtx struct gerekir — bu PR'ın
kapsamı dışı (minimal soyutlama kuralı).

## Public API değişmedi
- `connection::handle(socket, peer)` imzası aynı (server.rs çağrısı)
- `sender.rs`'in kullandığı `build_connection_response_accept`,
  `build_paired_key_*`, `build_sharing_cancel`, `send_sharing_frame`,
  `send_disconnection` → `mod.rs`'te `pub(crate) use frames::{...}`
  ile re-export, caller'lar hiç değişmedi.
- `main.rs` `mod connection;` dokunulmadı (cargo mod.rs'i otomatik
  çözüyor).
- `lib.rs` — connection expose edilmiyor (integration testlerden
  erişim yok, değişiklik gerekmedi).

## Test
- 397 test, 0 fail — tüm PR #1-#3 regression suite'i geçiyor.
- Mevcut unit testler alt-modüllere taşındı (`super::*` ile lokal
  erişim): sanitize.rs'e sanitize testleri, sharing.rs'e drain/
  privacy testleri, request.rs'e parse_remote_name testleri.

## Açık kalan
- `handle()` 332 satır — encrypted loop body extraction sonraki PR
  adayı (ConnectionCtx struct + run_encrypted_loop()).
- consent.rs 368 satır — Introduction branch 250 satır; bölme
  mantıksal akışı dağıtırdı.
